requirements.txt
# customer_portal/requirements.txt
# Core Framework
Flask==2.3.2
Waitress==2.1.2

# Database
pyodbc==4.0.39

# Environment
python-dotenv==1.0.0

# Security / Hashing / Templates
Werkzeug==2.3.6
MarkupSafe==2.1.3
Jinja2==3.1.2
cryptography==41.0.4 # Werkzeug dependency

# Excel Export
openpyxl==3.1.2